Moved addEvent responsibility to EventManager
